develop: renamed 'containers' directory to 'dataStructures'; continued working on jsdoc updates
