+++ index.js
@@ -1,8 +1,8 @@
 require('babel-polyfill');
 
 window.pjs = {
-    monads: require('./src/containers/monads/monads'),
-    functors: require('./src/containers/functors/functors'),
+    monads: require('./src/dataStructures/monads/monads'),
+    functors: require('./src/dataStructures/functors/functors'),
     groups: require('./groups/groups'),
     stream: require('./src/streams/observable'),
     combinators: require('./src/combinators'),
