+++ src/dataStructures/functors/constant_functor.js
@@ -1,4 +1,4 @@
-import { equalMaker, pointMaker, stringMaker, valueOf, get, orElse, getOrElse } from '../containerHelpers';
+import { equalMaker, pointMaker, stringMaker, valueOf, get, orElse, getOrElse } from '../dataStructureHelpers';
 
 /**
  * @sig
