+++ src/dataStructures/functors/list_functor.js
