+++ test/unit/containers/list_iterators.spec.js
@@ -1,8 +1,8 @@
 import { all, any, except, intersect, union, map, chain, groupBy, sortBy, prepend, concat, groupJoin, join, zip, filter, intersperse,
         contains, first, last, count, foldLeft, reduceRight, distinct, ofType, binarySearch, equals, takeWhile, skipWhile, reverse,
-        copyWithin, fill, findIndex, findLastIndex, repeat, foldRight, unfold } from '../../../src/containers/list_iterators';
-import { createListCreator } from '../../../src/containers/list_helpers';
-import { list_functor, ordered_list_functor } from '../../../src/containers/functors/list_functor';
+        copyWithin, fill, findIndex, findLastIndex, repeat, foldRight, unfold } from '../../../src/dataStructures/list_iterators';
+import { createListCreator } from '../../../src/dataStructures/list_helpers';
+import { list_functor, ordered_list_functor } from '../../../src/dataStructures/functors/list_functor';
 import { cacher, javaScriptTypes, sortDirection, typeNames } from '../../../src/helpers';
 import { testData } from '../../testData';
